:recycle: Refactor transition stuff
